Изменил согласно замечаниям

+++ src/main/java/faang/school/godbless/calculator/Main.java
@@ -5,10 +5,10 @@ import java.util.List;
 
 public class Main {
     public static void main(String[] args) {
-        List<Integer> numbers = Arrays.asList(1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
+        List<Long> numbers = Arrays.asList(1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L);
 
-        Integer sumResult = CalculatorUtil.sum(numbers);
-        Integer productResult = CalculatorUtil.product(numbers);
+        Long sumResult = CalculatorUtil.sum(numbers);
+        Long productResult = CalculatorUtil.product(numbers);
 
         System.out.printf("The sum of the numbers is: %d\n", sumResult);
         System.out.printf("The product of the numbers is: %d\n", productResult);
